moved NaubFactory in own file
